+++ enterprise/cmd/frontend/internal/context/resolvers/context_test.go
@@ -139,15 +139,17 @@ func TestContextResolver(t *testing.T) {
 		contextClient,
 	)
 
+	truePtr := true
 	conf.Mock(&conf.Unified{
 		SiteConfiguration: schema.SiteConfiguration{
+			CodyEnabled: &truePtr,
 			Embeddings:  &schema.Embeddings{Enabled: true},
-			Completions: &schema.Completions{Enabled: true},
+			Completions: &schema.Completions{},
 		},
 	})
 
 	ctx = actor.WithActor(ctx, actor.FromMockUser(1))
-	ffs := featureflag.NewMemoryStore(map[string]bool{"cody-experimental": true}, nil, nil)
+	ffs := featureflag.NewMemoryStore(map[string]bool{"cody": true}, nil, nil)
 	ctx = featureflag.WithFlags(ctx, ffs)
 
 	results, err := resolver.GetCodyContext(ctx, graphqlbackend.GetContextArgs{